GUI Update 1/5: admin
git-svn-id: https://i-mscp.svn.sourceforge.net/svnroot/i-mscp/trunk@1243 15a38060-5cfe-4f5c-bf9d-b5d42b1378a7

+++ gui/admin/dhp.php
@@ -1,45 +1,57 @@
 <?php
 /**
- *  ispCP (OMEGA) - Virtual Hosting Control System | Omega Version
+ * ispCP ω (OMEGA) a Virtual Hosting Control System
  *
- *  @copyright 	2001-2006 by moleSoftware GmbH
- *  @copyright 	2006-2007 by ispCP | http://isp-control.net
- *  @link 		http://isp-control.net
- *  @author		ispCP Team (2007)
+ * @copyright 	2001-2006 by moleSoftware GmbH
+ * @copyright 	2006-2008 by ispCP | http://isp-control.net
+ * @version 	SVN: $ID$
+ * @link 		http://isp-control.net
+ * @author 		ispCP Team
  *
- *  @license
- *  This program is free software; you can redistribute it and/or modify it under
- *  the terms of the MPL General Public License as published by the Free Software
- *  Foundation; either version 1.1 of the License, or (at your option) any later
- *  version.
- *  You should have received a copy of the MPL Mozilla Public License along with
- *  this program; if not, write to the Open Source Initiative (OSI)
- *  http://opensource.org | osi@opensource.org
- **/
+ * @license
+ *   This program is free software; you can redistribute it and/or modify it under
+ *   the terms of the MPL General Public License as published by the Free Software
+ *   Foundation; either version 1.1 of the License, or (at your option) any later
+ *   version.
+ *   You should have received a copy of the MPL Mozilla Public License along with
+ *   this program; if not, write to the Open Source Initiative (OSI)
+ *   http://opensource.org | osi@opensource.org
+ */
 
 require '../include/ispcp-lib.php';
 
 check_login(__FILE__);
 
-$theme_color = Config::get('USER_INITIAL_THEME');
+if (strtolower(Config::get('HOSTING_PLANS_LEVEL')) != 'admin') {
+	header('Location: index.php');
+	die();
+}
+
 
 if(isset($_GET['hpid']) && is_numeric($_GET['hpid']))
 	$hpid = $_GET['hpid'];
-else{
+else {
 	$_SESSION['hp_deleted'] = '_no_';
 	Header('Location: hp.php');
 	die();
 }
 
+// Check if there is no order for this plan
+$res = exec_query($sql, "SELECT COUNT(id) FROM `orders` WHERE `plan_id`=? AND `status`='new'", array($hpid));
+$data = $res->FetchRow();
+if ($data['0'] > 0) {
+	$_SESSION['hp_deleted_ordererror'] = '_yes_';
+	header("Location: hp.php");
+	die();
+}
+
 // Try to delete hosting plan from db
-//
-$query = 'delete from hosting_plans where id=?';
+$query = 'delete from hosting_plans where id = ?';
 $res = exec_query($sql, $query, array($hpid));
 
 $_SESSION['hp_deleted'] = '_yes_';
 
-Header('Location: hp.php');
-
+header('Location: hp.php');
 die();
 
 ?>
